refactor: Refactor messages components
