+++ src/utils/rooms.ts
@@ -251,6 +251,7 @@ export const handleRoomEvents = async (
       allMessageProperties.push({
         kind: MessageKind.Unread,
         data: {lastReadEventId},
+        messageId: lastReadEventId,
       })
     }
 
@@ -287,6 +288,7 @@ const processPatterns = (
   ) {
     return {
       kind: MessageKind.EventGroup,
+      messageId: lastMessage.eventId,
       data: {
         eventMessages: [lastMessage, currentMessage],
         eventGroupMainBody: {
@@ -301,6 +303,7 @@ const processPatterns = (
 
   return {
     kind: MessageKind.EventGroup,
+    messageId: lastMessage.eventId,
     data: {
       eventMessages: [lastMessage, currentMessage],
       eventGroupMainBody: {
@@ -321,6 +324,7 @@ const updateEventGroup = (
 
   const partialMessage: AnyMessage = {
     kind: MessageKind.EventGroup,
+    messageId: newEvent.eventId,
     data: {
       eventMessages: [...eventGroup.eventMessages, newEvent],
       eventGroupMainBody: eventGroup.eventGroupMainBody,
@@ -414,6 +418,7 @@ export const handleRoomMessageEvent = async (
 
   return {
     kind: MessageKind.Event,
+    messageId: event.event.event_id,
     data: {
       eventId: event.event.event_id,
       timestamp: event.localTimestamp,
@@ -834,8 +839,10 @@ export const handleMessage = async (
           validateReplyMessage(eventContent.body)
         ) {
           const replyData = parseReplyMessageFromBody(eventContent.body)
+
           return {
             kind: MessageKind.Reply,
+            messageId: messageBaseProperties.messageId,
             data: {
               ...messageBaseProperties,
               text: replyData.message,
@@ -849,6 +856,7 @@ export const handleMessage = async (
 
       return {
         kind: MessageKind.Text,
+        messageId: messageBaseProperties.messageId,
         data: {
           ...messageBaseProperties,
           text: eventContent.body,
@@ -865,6 +873,7 @@ export const handleMessage = async (
 
       return {
         kind: MessageKind.Image,
+        messageId: messageBaseProperties.messageId,
         data: {
           ...messageBaseProperties,
           imageUrl: getImageUrl(eventContent.url, room.client),
@@ -883,6 +892,7 @@ export const handleMessage = async (
 
       return {
         kind: MessageKind.File,
+        messageId: messageBaseProperties.messageId,
         data: {
           ...messageBaseProperties,
           fileUrl: getFileUrl(fileUrl, room.client),
@@ -901,6 +911,7 @@ export const handleMessage = async (
 
       return {
         kind: MessageKind.Audio,
+        messageId: messageBaseProperties.messageId,
         data: {
           ...messageBaseProperties,
           audioUrl: getFileUrl(audioUrl, room.client),
@@ -927,6 +938,7 @@ export const handleMessage = async (
 
       return {
         kind: MessageKind.Video,
+        messageId: messageBaseProperties.messageId,
         data: {
           ...messageBaseProperties,
           url: videoUrl,
@@ -974,6 +986,7 @@ const convertToMessageDeleted = (
 
   return {
     kind: MessageKind.Text,
+    messageId: eventId,
     data: {
       userId: sender.userId,
       authorAvatarUrl: getImageUrl(sender.getMxcAvatarUrl(), room.client),
